+++ sismic/interpreter/__init__.py
@@ -1,0 +1,5 @@
+from . import helpers
+from .interpreter import Interpreter
+
+__all__ = ['Interpreter']
+