+++ tests/test_interpreter.py
@@ -1,6 +1,7 @@
 import unittest
 from sismic import io
-from sismic.interpreter import Interpreter, run_in_background, log_trace
+from sismic.interpreter import Interpreter
+from sismic.interpreter.helpers import run_in_background, log_trace
 from sismic import exceptions
 from sismic.code import DummyEvaluator
 from sismic.model import Event, InternalEvent