+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ Unreleased
 - (Added) ``interpreter.raise_event`` to send events from within the statechart.
 - (Changed) Events sent by a code evaluator are now returned by the ``execute_*`` methods instead of being
   automatically added to the interpreter's queue.
+- (Changed) Moved ``run_in_background`` and ``log_trace`` from ``sismic.interpreter`` to the newly added
+  ``sismic.interpreter.helpers``.
 
 0.21.0 (2016-04-22)
 -------------------